// eshop-web/src/main/resources/client/src/app/core/enums/battery-search-key.enum.ts
export enum BATTERY_SEARCH_KEY {
    VOLTAGE = 'voltage',
    AMPERE_HOUR = 'ampereHour',
    WIDTH = 'width',
    HEIGHT = 'height',
    LENGTH = 'length',
    INTERCONNECTION = 'interconnection',
    TYPE_OF_POLE = 'typeOfPole',
    WITHOUT_START_STOP = 'withoutStartStop',
    WITH_START_STOP = 'withStartStop',
    IS_RANGE_AMPERE_HOUR_CB = 'isRangeAmpereHourCb',
    IS_RANGE_LENGTH_CB = 'isRangeLengthCb',
    IS_RANGE_WIDTH_CB = 'isRangeWidthCb',
    IS_RANGE_HEIGHT_CB = 'isRangeHeightCb'
}